@@ -23,7 +23,7 @@
     "require": {
         "php": " >=8.3",
         "ext-json": "*",
-        "babdev/pagerfanta-bundle": "^3.8",
+        "babdev/pagerfanta-bundle": "^4.5",
         "ibexa/content-forms": "~5.0.x-dev",
         "ibexa/core": "~5.0.x-dev",
         "ibexa/design-engine": "~5.0.x-dev",
@@ -34,24 +34,24 @@
         "jms/translation-bundle": "^2.4",
         "knplabs/knp-menu-bundle": "^3.0",
         "mck89/peast": "^1.9",
-        "symfony/asset": "^6.4",
-        "symfony/config": "^6.4",
-        "symfony/console": "^6.4",
-        "symfony/dependency-injection": "^6.4",
+        "symfony/asset": "^7.2",
+        "symfony/config": "^7.2",
+        "symfony/console": "^7.2",
+        "symfony/dependency-injection": "^7.2",
         "symfony/deprecation-contracts": "^2.5 || ^3.0",
-        "symfony/event-dispatcher": "^6.4",
-        "symfony/filesystem": "^6.4",
-        "symfony/form": "^6.4",
-        "symfony/http-foundation": "^6.4",
-        "symfony/http-kernel": "^6.4",
-        "symfony/options-resolver": "^6.4",
-        "symfony/routing": "^6.4",
-        "symfony/security-core": "^6.4",
-        "symfony/security-http": "^6.4",
-        "symfony/translation": "^6.4",
-        "symfony/validator": "^6.4",
-        "symfony/webpack-encore-bundle": "^1.17",
-        "symfony/yaml": "^6.4",
+        "symfony/event-dispatcher": "^7.2",
+        "symfony/filesystem": "^7.2",
+        "symfony/form": "^7.2",
+        "symfony/http-foundation": "^7.2",
+        "symfony/http-kernel": "^7.2",
+        "symfony/options-resolver": "^7.2",
+        "symfony/routing": "^7.2",
+        "symfony/security-core": "^7.2",
+        "symfony/security-http": "^7.2",
+        "symfony/translation": "^7.2",
+        "symfony/validator": "^7.2",
+        "symfony/webpack-encore-bundle": "^2.2",
+        "symfony/yaml": "^7.2",
         "twig/intl-extra": "^3.0",
         "twig/string-extra": "^3.0",
         "twig/twig": "^3.0",
@@ -59,16 +59,16 @@
     },
     "require-dev": {
         "ext-zip": "*",
-        "dama/doctrine-test-bundle": "^v6.7",
+        "dama/doctrine-test-bundle": "^8.2",
         "ibexa/behat": "~5.0.x-dev",
         "ibexa/code-style": "~2.0.0",
         "ibexa/doctrine-schema": "~5.0.x-dev",
         "ibexa/http-cache": "~5.0.x-dev",
         "ibexa/notifications": "~5.0.x-dev",
+        "ibexa/rector": "~5.0.x-dev",
         "ibexa/test-core": "~5.0.x-dev",
         "ibexa/test-rest": "~5.0.x-dev",
-        "ibexa/rector": "~5.0.x-dev",
-        "matthiasnoback/symfony-dependency-injection-test": "^4.0",
+        "matthiasnoback/symfony-dependency-injection-test": "^5.0",
         "phpstan/phpstan": "^2.0",
         "phpstan/phpstan-phpunit": "^2.0",
         "phpstan/phpstan-symfony": "^2.0",
