@@ -7,6 +7,7 @@
 
 namespace Ibexa\AdminUi\Form\DataTransformer\FieldType;
 
+use Ibexa\Core\FieldType\BinaryFile\Value;
 use Symfony\Component\Form\DataTransformerInterface;
 
 /**
@@ -16,12 +17,7 @@ use Symfony\Component\Form\DataTransformerInterface;
  */
 class BinaryFileValueTransformer extends AbstractBinaryBaseTransformer implements DataTransformerInterface
 {
-    /**
-     * @param \Ibexa\Core\FieldType\BinaryFile\Value $value
-     *
-     * @return array
-     */
-    public function transform($value)
+    public function transform(mixed $value): array
     {
         if (null === $value) {
             $value = $this->fieldType->getEmptyValue();
@@ -33,14 +29,7 @@ class BinaryFileValueTransformer extends AbstractBinaryBaseTransformer implement
         );
     }
 
-    /**
-     * @param array $value
-     *
-     * @return \Ibexa\Core\FieldType\BinaryFile\Value
-     *
-     * @throws \Symfony\Component\Form\Exception\TransformationFailedException
-     */
-    public function reverseTransform($value)
+    public function reverseTransform(mixed $value): Value
     {
         /** @var \Ibexa\Core\FieldType\BinaryFile\Value $valueObject */
         $valueObject = $this->getReverseTransformedValue($value);
