@@ -32,7 +32,7 @@ class FieldValueTransformer implements DataTransformerInterface
      *
      * @return array|null the value's hash, or null if $value was not a FieldType Value
      */
-    public function transform($value)
+    public function transform(mixed $value): ?array
     {
         if (!$value instanceof Value) {
             return null;
@@ -49,7 +49,7 @@ class FieldValueTransformer implements DataTransformerInterface
      *
      * @return \Ibexa\Contracts\Core\FieldType\Value
      */
-    public function reverseTransform($value)
+    public function reverseTransform(mixed $value): ?Value
     {
         if ($value === null) {
             return $this->fieldType->getEmptyValue();
