@@ -7,6 +7,7 @@
 
 namespace Ibexa\AdminUi\Form\DataTransformer\FieldType;
 
+use Ibexa\Core\FieldType\Image\Value;
 use Symfony\Component\Form\DataTransformerInterface;
 
 /**
@@ -21,7 +22,7 @@ class ImageValueTransformer extends AbstractBinaryBaseTransformer implements Dat
      *
      * @return array
      */
-    public function transform($value)
+    public function transform(mixed $value): ?array
     {
         if (null === $value) {
             $value = $this->fieldType->getEmptyValue();
@@ -40,7 +41,7 @@ class ImageValueTransformer extends AbstractBinaryBaseTransformer implements Dat
      *
      * @throws \Symfony\Component\Form\Exception\TransformationFailedException
      */
-    public function reverseTransform($value)
+    public function reverseTransform(mixed $value): ?Value
     {
         /** @var \Ibexa\Core\FieldType\Image\Value $valueObject */
         $valueObject = $this->getReverseTransformedValue($value);
