@@ -7,28 +7,30 @@
 
 namespace Ibexa\AdminUi\Form\DataTransformer\FieldType;
 
+use Ibexa\Core\FieldType\Media\Value;
 use Symfony\Component\Form\DataTransformerInterface;
 
 /**
  * Data transformer for ezmedia field type.
  *
- * {@inheritdoc}
+ * @phpstan-type TMediaProperties = array{hasController: bool, loop: bool, autoplay: bool, width: int, height: int}
+ *
+ * @phpstan-implements \Symfony\Component\Form\DataTransformerInterface<\Ibexa\Core\FieldType\Media\Value, TMediaProperties>
  */
 class MediaValueTransformer extends AbstractBinaryBaseTransformer implements DataTransformerInterface
 {
     /**
-     * @param \Ibexa\Core\FieldType\Media\Value $value
-     *
-     * @return array
+     * @phpstan-return TMediaProperties
      */
-    public function transform($value)
+    public function transform(mixed $value): array
     {
         if (null === $value) {
             $value = $this->fieldType->getEmptyValue();
         }
 
+        /** @phpstan-var TMediaProperties */
         return array_merge(
-            $this->getDefaultProperties(),
+            $this->getDefaultProperties() ?? [],
             [
                 'hasController' => $value->hasController,
                 'loop' => $value->loop,
@@ -40,13 +42,9 @@ class MediaValueTransformer extends AbstractBinaryBaseTransformer implements Dat
     }
 
     /**
-     * @param array $value
-     *
-     * @return \Ibexa\Core\FieldType\Media\Value
-     *
-     * @throws \Symfony\Component\Form\Exception\TransformationFailedException
+     * @phpstan-param TMediaProperties $value
      */
-    public function reverseTransform($value)
+    public function reverseTransform(mixed $value): Value
     {
         /** @var \Ibexa\Core\FieldType\Media\Value $valueObject */
         $valueObject = $this->getReverseTransformedValue($value);
